Added few more fields and improved user authentication

+++ app/Http/Controllers/ItemController.php
@@ -10,6 +10,7 @@ use \Session;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Item;
+use Auth;
 
 
 
@@ -27,11 +28,19 @@ class ItemController extends Controller
             'items' => $items
         ]);*/
 
-        // get all the items
-        $items = Item::all();
+        if(Auth::check()) {
+            if (Auth::user()) {
+                $items = Item::where('user_id', Auth::id())
+                    ->get();
+        } else {
+                // get all the items
+                $items = Item::all();
+            }
 
-        // load the view and pass the items
-        return view('items.index')->with('items', $items);
+
+            // load the view and pass the items
+            return view('items.index')->with('items', $items);
+        }
     }
 
     /**
@@ -52,26 +61,33 @@ class ItemController extends Controller
      */
     public function store(Request $request)
     {
-        // validate
-        $rules = array(
-            'description' => 'required'
-        );
-        $validator = Validator::make($request->all(), $rules);
+        if(Auth::check()) {
+            // validate
+            $rules = array(
+                'description' => 'required',
+                'quantity' => 'required',
+                'price' => 'required'
+            );
+            $validator = Validator::make($request->all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return redirect('items/create')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
-        } else {
-            // store
-            $item = new Item;
-            $item->description = $request->description;
-            $item->save();
+            // process the login
+            if ($validator->fails()) {
+                return redirect('items/create')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $item = new Item;
+                $item->description = $request->description;
+                $item->quantity = $request->quantity;
+                $item->price = $request->price;
+                $item->user_id = Auth::id();
+                $item->save();
 
-            // redirect
-            Session::flash('message', 'Successfully added new item!');
-            return redirect('/items');
+                // redirect
+                Session::flash('message', 'Successfully added new item!');
+                return redirect('/items');
+            }
         }
     }
 
@@ -114,26 +130,30 @@ class ItemController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // validate
-        $rules = array(
-            'description' => 'required'
-        );
-        $validator = Validator::make($request->all(), $rules);
+        if(Auth::check()) {
+            // validate
+            $rules = array(
+                'description' => 'required'
+            );
+            $validator = Validator::make($request->all(), $rules);
 
-        // process the login
-        if ($validator->fails()) {
-            return Redirect::to('items/' . $id . '/edit')
-                ->withErrors($validator)
-                ->withInput(Input::except('password'));
-        } else {
-            // store
-            $item = Item::find($id);
-            $item->description = $request->description;
-            $item->save();
+            // process the login
+            if ($validator->fails()) {
+                return Redirect::to('items/' . $id . '/edit')
+                    ->withErrors($validator)
+                    ->withInput(Input::except('password'));
+            } else {
+                // store
+                $item = Item::find($id);
+                $item->description = $request->description;
+                $item->quantity = $request->quantity;
+                $item->price = $request->price;
+                $item->save();
 
-            // redirect
-            Session::flash('message', 'Successfully updated item!');
-            return redirect('/items');
+                // redirect
+                Session::flash('message', 'Successfully updated item!');
+                return redirect('/items');
+            }
         }
     }
 
@@ -145,12 +165,14 @@ class ItemController extends Controller
      */
     public function destroy($id)
     {
-        // delete
-        $item = Item::find($id);
-        $item->delete();
+        if(Auth::check()) {
+            // delete
+            $item = Item::find($id);
+            $item->delete();
 
-        // redirect
-        Session::flash('message', 'Successfully deleted the item!');
-        return redirect('/items');
+            // redirect
+            Session::flash('message', 'Successfully deleted the item!');
+            return redirect('/items');
+        }
     }
 }
